+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # COBOL Source colouriser for Visual Studio Code
 
+## Version - 3.2.0
+- Merge goto definition changes (Thanks Ted)
+- Move away from using workspace to workspace folders
+- fix schema warning message
+
 ## Version - 3.1.0
 - Fix issue #13 - Some lines are incorrectly coloured
 - Fix firstLine regex